+++ lib/digicert/cli/order.rb
@@ -3,14 +3,7 @@ require "digicert/cli/order_reissuer"
 
 module Digicert
   module CLI
-    class Order
-      attr_reader :order_id, :options
-
-      def initialize(attributes = {})
-        @options = attributes
-        @order_id = options.delete(:order_id)
-      end
-
+    class Order < Digicert::CLI::Base
       def list
         filtered_orders = apply_filters(orders)
         display_orders_in_table(filtered_orders)
@@ -25,6 +18,17 @@ module Digicert
         Digicert::CLI::OrderReissuer.new(
           order_id: order_id, **options
         ).create
+      end
+
+      def self.local_options
+        [
+          ["-q", "--quiet",  "Flag to return resource Id only"],
+          ["-s", "--status STATUS", "Use to specify the order status"],
+          ["-n", "--product_type NAME_ID", "The Digicert product name Id"],
+          ["-f", "--fetch", "Flag to fetch resource after certian operation"],
+          ["-r", "--crt CSR_FILE", "Full path for the csr file"],
+          ["-p", "--output DOWNLOAD_PATH", "Path to download the certificate"]
+        ]
       end
 
       private
